requirements.txt
requests==2.32.3
python-dotenv==1.0.1
pyyaml==6.0.3
playwright==1.58.0
mkdocs==1.6.1
mkdocs-material==9.6.14
pillow==12.1.1
